# backend/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/items": {
      "post": {
        "operationId": "ItemsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateItemDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Item created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new item",
        "tags": [
          "Items"
        ]
      },
      "get": {
        "operationId": "ItemsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Items retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ItemResponseDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all items",
        "tags": [
          "Items"
        ]
      }
    },
    "/items/bulk": {
      "post": {
        "operationId": "ItemsController_createBulk",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkCreateItemDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Items created successfully"
          },
          "400": {
            "description": "Invalid input data"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create multiple items at once",
        "tags": [
          "Items"
        ]
      }
    },
    "/items/inventory-value": {
      "get": {
        "operationId": "ItemsController_getInventoryValue",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Inventory value calculated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "totalValue": {
                      "type": "number",
                      "description": "Total inventory value"
                    },
                    "itemsWithValue": {
                      "type": "number",
                      "description": "Number of items with sellprice"
                    },
                    "totalItems": {
                      "type": "number",
                      "description": "Total number of items"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get total inventory value based on selling prices",
        "tags": [
          "Items"
        ]
      }
    },
    "/items/statistics/status": {
      "get": {
        "operationId": "ItemsController_getStatusStatistics",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Status distribution statistics retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "status": {
                            "type": "string",
                            "example": "Good"
                          },
                          "count": {
                            "type": "number",
                            "example": 15
                          }
                        }
                      }
                    },
                    "total": {
                      "type": "number",
                      "example": 50
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get items status distribution statistics",
        "tags": [
          "Items"
        ]
      }
    },
    "/items/search": {
      "get": {
        "operationId": "ItemsController_search",
        "parameters": [
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Search term",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Search results retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ItemResponseDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Search items",
        "tags": [
          "Items"
        ]
      }
    },
    "/items/{id}": {
      "get": {
        "operationId": "ItemsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Item ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Item retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Item not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get item by ID",
        "tags": [
          "Items"
        ]
      },
      "patch": {
        "operationId": "ItemsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Item ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateItemDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Item updated successfully"
          },
          "400": {
            "description": "Invalid input data"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Item not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update item",
        "tags": [
          "Items"
        ]
      },
      "delete": {
        "operationId": "ItemsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Item ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Item deleted successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Item not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete item",
        "tags": [
          "Items"
        ]
      }
    },
...
    "/favourites/item/{itemId}": {
      "post": {
        "operationId": "FavouritesController_createByItem",
        "parameters": [
          {
            "name": "itemId",
            "required": true,
            "in": "path",
            "description": "Item ID to add to favourites",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Item added to favourites successfully"
          },
          "400": {
            "description": "Item already in favourites"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Item not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Add item to favourites",
        "tags": [
          "Favourites"
        ]
      },
      "delete": {
        "operationId": "FavouritesController_removeByItem",
        "parameters": [
          {
            "name": "itemId",
            "required": true,
            "in": "path",
            "description": "Item ID to remove from favourites",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Item removed from favourites successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Favourite not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Remove item from favourites",
        "tags": [
          "Favourites"
        ]
      }
    },
...
    "/projects/{id}/items/{itemId}": {
      "patch": {
        "description": "Modifies the quantity or active state of an item in a project",
        "operationId": "ProjectsController_updateProjectItem",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Project ID",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "itemId",
            "required": true,
            "in": "path",
            "description": "Item ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProjectItemDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Project item updated successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Project or item not found in the project"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Modify an item in a project",
        "tags": [
          "Projects"
        ]
      },
      "delete": {
        "description": "Removes an item completely from a project",
        "operationId": "ProjectsController_removeItemFromProject",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Project ID",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "itemId",
            "required": true,
            "in": "path",
            "description": "Item ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Item removed from project successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Project or item not found in the project"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Remove an item from a project",
        "tags": [
          "Projects"
        ]
      }
    },
...
  },
  "info": {
    "title": "ShelfSpot API",
    "description": "API for ShelfSpot application with JWT Authentication",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
...
      "CreateItemDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Laptop Dell XPS 13",
            "description": "Item name"
          },
          "quantity": {
            "type": "number",
            "example": 1,
            "description": "Item quantity",
            "minimum": 1
          },
          "status": {
            "type": "string",
            "example": "Available",
            "description": "Item status"
          },
          "itemLink": {
            "type": "string",
            "example": "https://example.com/item-link",
            "description": "Link to the item"
          },
          "price": {
            "type": "number",
            "example": 999.99,
            "description": "Purchase price of the item"
          },
          "sellprice": {
            "type": "number",
            "example": 899.99,
            "description": "Selling price of the item"
          },
          "consumable": {
            "type": "boolean",
            "example": true,
            "description": "Indicates if the item is a consumable"
          },
          "roomId": {
            "type": "number",
            "example": 1,
            "description": "Room ID"
          },
          "placeId": {
            "type": "number",
            "example": 1,
            "description": "Place ID"
          },
          "containerId": {
            "type": "number",
            "example": 1,
            "description": "Container ID"
          }
        },
        "required": [
          "name",
          "quantity",
          "roomId"
        ]
      },
      "ItemResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1,
            "description": "Item ID"
          },
          "name": {
            "type": "string",
            "example": "Laptop Dell XPS 13",
            "description": "Item name"
          },
          "quantity": {
            "type": "number",
            "example": 1,
            "description": "Item quantity"
          },
          "status": {
            "type": "string",
            "example": "Available",
            "description": "Item status"
          },
          "itemLink": {
            "type": "string",
            "example": "https://example.com/item-link",
            "description": "Link to the item"
          },
          "consumable": {
            "type": "boolean",
            "example": true,
            "description": "Indicates if the item is a consumable"
          },
          "price": {
            "type": "number",
            "example": 999.99,
            "description": "Price of the item"
          },
          "createdAt": {
            "type": "string",
            "example": "2024-01-01T00:00:00.000Z",
            "description": "Creation date"
          },
          "updatedAt": {
            "type": "string",
            "example": "2024-01-01T00:00:00.000Z",
            "description": "Last update date"
          },
          "room": {
            "type": "object",
            "example": {
              "id": 1,
              "name": "Living Room"
            },
            "description": "Room information"
          },
          "place": {
            "type": "object",
            "example": {
              "id": 1,
              "name": "Kitchen Counter"
            },
            "description": "Place information"
          },
          "container": {
            "type": "object",
            "example": {
              "id": 1,
              "name": "Storage Box"
            },
            "description": "Container information"
          }
        },
        "required": [
          "id",
          "name",
          "quantity",
          "status",
          "createdAt",
          "updatedAt",
          "room",
          "place"
        ]
      },
      "BulkCreateItemDto": {
        "type": "object",
        "properties": {
          "items": {
            "description": "Array of items to be created",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateItemDto"
            }
          }
        },
        "required": [
          "items"
        ]
      },
      "UpdateItemDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Laptop Dell XPS 13",
            "description": "Item name"
          },
          "quantity": {
            "type": "number",
            "example": 1,
            "description": "Item quantity",
            "minimum": 1
          },
          "status": {
            "type": "string",
            "example": "Available",
            "description": "Item status"
          },
          "itemLink": {
            "type": "string",
            "example": "https://example.com/item-link",
            "description": "Link to the item"
          },
          "price": {
            "type": "number",
            "example": 999.99,
            "description": "Purchase price of the item"
          },
          "sellprice": {
            "type": "number",
            "example": 899.99,
            "description": "Selling price of the item"
          },
          "roomId": {
            "type": "number",
            "example": 1,
            "description": "Room ID"
          },
          "placeId": {
            "type": "number",
            "example": 1,
            "description": "Place ID"
          },
          "containerId": {
            "type": "number",
            "example": 1,
            "description": "Container ID"
          },
          "tags": {
            "example": [
              "electronics",
              "work"
            ],
            "description": "Array of tag names",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "UpdateProjectItemDto": {
        "type": "object",
        "properties": {
          "quantity": {
            "type": "number",
            "example": 5,
            "description": "New quantity of this item used in the project",
            "minimum": 1
          },
          "isActive": {
            "type": "boolean",
            "example": false,
            "description": "Whether the usage of this item is active in the project"
          }
        }
      }
    }
  }
}